:sparkles: add list directory and model file
