+++ lib/main/home/home.dart
@@ -1,6 +1,6 @@
 import 'package:Pilll/main/record/pill_sheet.dart';
 import 'package:Pilll/main/record/record_taken_information.dart';
-import 'package:Pilll/settings/settings.dart';
+import 'package:Pilll/settings/list/settings.dart';
 import 'package:Pilll/theme/color.dart';
 import 'package:Pilll/theme/text_color.dart';
 import 'package:flutter/material.dart';
